Added new doc on working with github

+++ scripts/generate_mandeldocs.sh
@@ -51,25 +51,7 @@ GenMandelToolDoc() {
   mkdir markdown_out
   cp -R docs/* markdown_out/
   # general_info does not get copied to developer-tools
-  # will be copied over later in CopyTutorialsAndGuides()
-  rm -rf markdown_out/general_info
+  [ -d markdown_out/general_info ] && rm -rf markdown_out/general_info
 
   cp -R markdown_out/* $DOC_ROOT
-}
-
-CopyTutorialsAndGuides() {
-  if [[ $# -lt 2 ]] ; then
-      echo 'NOT ENOUGH ARGS: specify web root,  specify script dir '
-      exit 1
-  fi
-  DOC_ROOT="${1}/devdocs/eosdocs/tutorials"
-  WORKING_DIR="${2}/../working"
-
-  # copy over glossary, guides, and tutorials
-  if [ -d ${WORKING_DIR}/mandel ]; then
-    cd ${WORKING_DIR}/mandel
-    cp tutorials/bios-boot-tutorial/README.md "${DOC_ROOT}/bios-boot-tutorial.md"
-    cp docs/general_info/glossary.md "${ROOT_DIR}/devdocs/eosdocs/general_info/"
-    cp docs/general_info/protocol-guides/* "${ROOT_DIR}/devdocs/eosdocs/general_info/protocol-guides/"
-  fi
 }
